_extensions/airtable-sync/requirements.txt
annotated-types==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
idna==3.10
inflection==0.5.1
pyairtable==3.0.2
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.3
